90% API

+++ src/main/java/com/project/retailhub/data/repository/ProductRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 import java.util.Optional;
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer>
+public interface ProductRepository extends JpaRepository<Product, Long>
 {
     Optional<Product> findByBarcode(String barcode);
     boolean existsByBarcode(String barcode);
